+++ app/src/main/java/dev/rikoapp/cleanphonelauncher/data/database/di/DatabaseModule.kt
@@ -3,6 +3,7 @@ package dev.rikoapp.cleanphonelauncher.data.database.di
 import androidx.room.Room
 import dev.rikoapp.cleanphonelauncher.data.RoomLocalFavoriteAppDataSource
 import dev.rikoapp.cleanphonelauncher.data.database.CleanPhoneLauncherDatabase
+import dev.rikoapp.cleanphonelauncher.data.database.MIGRATION_1_2
 import dev.rikoapp.cleanphonelauncher.domain.LocalFavoriteAppDataSource
 import org.koin.android.ext.koin.androidApplication
 import org.koin.core.module.dsl.singleOf
@@ -15,7 +16,7 @@ val databaseModule = module {
             androidApplication(),
             CleanPhoneLauncherDatabase::class.java,
             "CleanPhoneLauncher.db"
-        ).build()
+        ).addMigrations(MIGRATION_1_2).build()
     }
 
     single { get<CleanPhoneLauncherDatabase>().favoriteAppDao }