+++ app/src/main/java/dev/rikoapp/cleanphonelauncher/presentation/home/HomeScreenAction.kt
@@ -12,6 +12,7 @@ sealed class HomeScreenAction {
     data class OnFavoriteAppLongClick(val app: AppData) : HomeScreenAction()
     object OnFavoriteDialogDismiss : HomeScreenAction()
     data class OnRemoveFavorite(val packageName: String) : HomeScreenAction()
+    data class OnReorderFavorites(val orderedPackageNames: List<String>) : HomeScreenAction()
     data class OnAppInfoClick(val app: AppData) : HomeScreenAction()
     data class OnUninstallClick(val app: AppData) : HomeScreenAction()
     data class OnPhoneAppClick(val app: AppData) : HomeScreenAction()