+++ app/src/main/java/dev/rikoapp/cleanphonelauncher/presentation/ui/theme/Icon.kt
@@ -17,3 +17,7 @@ val CloseIcon: ImageVector
 val PhoneIcon: ImageVector
     @Composable
     get() = ImageVector.vectorResource(id = R.drawable.ic_phone)
+
+val DragHandleIcon: ImageVector
+    @Composable
+    get() = ImageVector.vectorResource(id = R.drawable.ic_drag_handle)